In the process of deubugging
I added the solution from MATLAB as the initial guess but the solver
still wouldn't go for it.

I also modified the code to do a loop instead of direct array
manipulation, mostly because of the if statement.

Upon examination by hand, it's clear why the program is encountering a
singular Jacobian: none of the derivatives depend on the x position, and
yet x is one of the state variables. I think I should get rid of it and
simply calculate x position from integration of Vx